osd/PG: _update_calc_stats: backfill targets do not always affect degraded
Our progress on a backfill target should only count toward degraded if
the target is needed in order for us to reach the target pool size.  If
we have remapped other complete copies, we can backfill and not be degraded
at all.  And we may be some combination of the two if there are multiple
backfill targets; use the target(s) with the most objects.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit d0213e3d37094185ceaac61553c401a0fc207112)