osd: send health-checks to mgr
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit f4b74125e44fe78154fb377fa06fc08b3325859d)

Conflicts:
	src/osd/OSD.cc
	src/osd/OSD.h: only the changes related to reporting
pending-creating-pgs are cherry-picked. because we want to minimize the
impact to luminous. and to remove the slow ops from cluster log is not
in the scope of this backport. also, with_unique_lock was introduced
after luminous is branched, so use the plain lock_guard<> instead.
	src/osd/OSDHealthMetric.h: because denc.h does not support
denc of enum types in luminous. so we need to case from/to osd_metric
when necessary.