mon/PGMap: 'unclean' does not imply damaged
Everything (that I can think of) that would lead to a PG being unclean is
already reported via another health message.  And there are cases where a
PG is unclean (e.g., because it is backfilling) where we are not degraded.

Fix by ignoring this flag in the health checks.

Signed-off-by: Sage Weil <sage@redhat.com>