Revert "osd/PG: allow local recovery reservations to be preempted"
This reverts commit ba106b6d32fb3138b7a61fe613b211e02f1e44aa.

There is another issue this brings to the surface with leaked recovery_ops.  A proper
fix for that will take more time to validate and it will be easier to re-backport this
patch as-is than to fix the other issue in a luminous-specific way (which will cause
a bunch of cherry-pick conflicts later).

Signed-off-by: Sage Weil <sage@redhat.com>