erasure-code: refactor the interfaces to hide internals from public
after the arraycode change, the old encode() and minimum_to_decode() are
actually replaced by the new counterparts, and are only used by the
ErasureCode itself. so they can be marked `private`, but the unittests
are still using them heavily. so they are kept `public`. and to avoid the
`-Woverloaded-virtual` warnings, because the new methods share the same
names with the old ones. a underscore is used as the prefix of the
internal methods.

Signed-off-by: Kefu Chai <kchai@redhat.com>