common/buffer: change atomic vars just once in crc32c
If the bufferlist has more than one element, we may be wasting time
modifying atomic variables. Reduce that to minimum by using local
variables and only in the end increasing atomic vars by larger
values. This commits also removes the

  if (buffer_track_crc) {
    (..)
  }

branches, as in production environments this is disabled anyway,
and increasing local variable is still way cheaper than branching,
predicted or not and taken or not.

Signed-off-by: Piotr Dałek <piotr.dalek@corp.ovh.com>