osd/PG: make pg versions non-overlapping in preparation for pg merge
In order to merge two PGs we'll need to make the pg log entries use
non-overlapping versions so that they can be merged together.  This is
only part of the solution: we make the versions non-contiguous and add
a debug option to do so randomly so we can ensure that we do not have
hidden assumptions around consecutive verion numbers.

Signed-off-by: Sage Weil <sage@redhat.com>