cmake: check gcc version not release date for libstdc++ saneness
there is chance that the release date of a minor or patch version of
libstdc++/gcc is *greater* than that of a major version. so this renders
the existing approach to check the __GLIBCPP__ useless. let's check the
gcc version instead. it's far from a perfect solution. but it's good
enough to cover most cases. assuming that most users use gcc with
libstdc++ comes with it. instead of using 1) gcc with a newer libstdc++,
or 2) clang with a old libstdc++.

Signed-off-by: Kefu Chai <kchai@redhat.com>