client: add delegation support for cephfs
Add the ability for ceph userland clients to request a delegation for a
cephfs open file. With this, userland will get a callback if there is a
request for access that conflicts with the delegation. The client is
then expected to return the delegation at which point the conflicting
access will succeed.

Handing out a delegation means that we're trusting the application to
give back the caps when they are needed elsewhere. If it fails to
uphold its end of the bargain, we need to take steps to forcibly revoke
them, so that other clients can make progress.

When that occurs, shut down the mount such that further calls return
-ENOTCONN. That should prevent the application from holding on to caps
indefinitely, and ensure that it will no longer do damage. It should
also prevent the client from ending up blacklisted.

Since this is an implicit agreement between ceph and the application,
we require that the application call ceph_set_deleg_timeout() to set
the delegation return timeout before we hand out any delegations. This
helps ensure that the various application and ceph cluster timeouts
are all in agreement.

An open call on ceph can easily return without any caps being granted to
the client. This is not generally a problem on ceph since it will wait
on caps before doing any operations on that open Fh, but it's less than
ideal when there is a delegation outstanding.

Both NFS and SMB servers can grant deny locks, and at least in the case
of NFS, those can be cached such that the server is unaware of them if
the client holds a delegation.

Most applications that use delegations will also want to ensure that
delegations are broken before allowing an open to proceed. A non-zero
delegation return timeout also cues the client to wait on a minimal set
of caps after an open return to ensure that this is the case.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit fad99776195e299b37071892e459ad8abf2fddfc)

Conflicts:
    src/client/Inode.cc (luminous does not have
        a5d97ad257d6fae8f66200513805e1778cb0c8ba so we preserve the following
        lines at the end of the Inode dtor:
          delete fcntl_locks;
          delete flock_locks;
        )