rgw: check read_op.read return value in RGWRados::copy_obj_data
The return-value-type of RGWRados::Object::Read::read is int, and
the return value of RGWRados::Object::Read::read is either the
length of object data which has been read if everything is OK,
or the error value which is negative value if something is wrong.

If an negative error value is assigned to read_len which is an
unsigned long long type, the read_len will be a very large value,
which will cause an error during RGWPutObjProcessor processing.

Signed-off-by: Enming Zhang <enming.zhang@umcloud.com>