mon/OSDMonitor: ensure lec only accounts for up osds
If we also consider down osds, we may very well be in a healthy state
but keeping maps as far back as the last epoch when a given osd went
down. If said osd stays down for eons, we will be keeping bajillions of
maps that we shouldn't.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>
(cherry picked from commit e62269c8929e414284ad0773c4a3c82e43735e4e)

Conflicts:
	src/mon/OSDMonitor.cc
          Lots of context changed, thus leading to conflicts.
          Cherry-pick was trivial once changes were put in the right
          place, and additional code removed.