osd,os,io: Initializing C_ProxyChunkRead members,queue,request
Fixes the coverity issue:

2. uninit_member: Non-static class member op_index is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member req_offset is not initialized
in this constructor nor in any functions that it calls.
CID 1424845 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
6. uninit_member: Non-static class member req_total_len is not
initialized in this constructor nor in any functions that it calls.

CID 1422461 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member queue is not initialized
in this constructor nor in any functions that it calls.

CID 1422362 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member request is not initialized
in this constructor nor in any functions that it calls

Signed-off-by: Amit Kumar amitkuma@redhat.com