radosgw: fix doubled underscore with s3/swift server-side copy
A name is almost an oid.  Except that if a name starts
with a _, the corresponding oid starts with a double _.
The copy logic passes the oid straight in as a name,
which results in a triple _ on the oid, resulting in
an object name with a double __.

Fix: remove one _ when converting an oid back into
a name, so that the final oid only has a double _.

Fixes: http://tracker.ceph.com/issues/22529

This change is not needed on master or Luminous, due to
already-merged #18662.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>