mgr: remove old 'config set' command
This is a cli command to set a config option on the mgr.  It doesn't
really belong because it only sets it on the mgr that happens to be
active.  It was probably mixed up with the asok or tell command.

Added in a101c4835944dec1b4c21d4c126d8605f29c8c29

Signed-off-by: Sage Weil <sage@redhat.com>