OSD: do not use deprecated random_shuffle
I'm a little concerned this might be slow to invoke on every pull. We may
want to move the random_device and mersenne twister up into a class instead
of constructing on every op, or use a cheaper method. But this is the proper
C++ solution for random integers now.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>