denc: Remove unneeded enable_ifs
The idiom
```
template<typename U = T>
static enable_if_t<!!sizeof(U) && ... >
```

Exists to force SFINAE dependent on some aspect of T on functions that
are themselves members of a template parameterized on T. Without some
other expression conjoined to it, this construct is just noise.

Also replace the reserve switch with constexpr if.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>