osd: only exit if *latest* map(s) say we are destroyed
It's possible our current map is older, we were destroyed then, but in
newer maps our osd was recreated.  This happens when the oldest map after
a recreated osd happens to land on an epoch where the osd was marked
destroyed.

Fix by only exiting if one of the newest maps says we are (still)
destroyed.

Fixes: http://tracker.ceph.com/issues/22673
Signed-off-by: Sage Weil <sage@redhat.com>