Allow the dmclock library to handle the "cost" of an operation/request. This is in contrast to just measuring ops. If the cost is set at the value 1, then it's equivalent to the old code.
The cost is calculated on the server side and affects the tag values
for the request. Furthermore, the cost is then sent back to the
client, so the ServiceTracker can use the cost to properly calculate
delta and rho values.

We now allow the delta and rho values sent with a request to be zero,
since the request cost is included on the server-side tag calculation,
and the request cost must be positive guaranteeing the advancement of
tags.

The OrigTracker has now been updated so as not to add one when
computing delta and rho.

The unit tests have been updated to use request costs. And the
simulator has been updated to use cost, and the cost of a client
group's requests can be configured.

The simulation has been changed to user the OrigTracker as the default
tracker.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>