Revised log append to rely on head/tail pointers in root object (not valid flags). Changed appender lock to normal Mutex, held while any thread is persisting the current batch of ready log entries.
Eliminated data block pool, and added a pointer in each log entry to an allocated
data buffer of the appropriate size. These are allocated with pmemobj_reserve()
when we need to copy the data out of the user buffer. These allocations are published
in the same transaction tht makes the log entries referring to them visible.

Among the things remaining TBD is deferring the write request when pmemobj_reserve()
fails (out of space).