build/ops: rpm: python3-ceph-argparse only if Python 2 available
The python3-ceph-argparse package is only necessary in environments where
Python 3 might not be available. For distros where Python 3 has become
standard with Python 2 an optional extension (instead of vice versa as
in the current status quo), the files in this package can (and should)
be included in ceph-common directly.

Signed-off-by: Nathan Cutler <ncutler@suse.com>