Enable 3rd party access to swift for most operations.
In order for cross-tenant swift acls to be useful, it's necessary to
enable "3rd party" access.  This means parsing the URL for the desired
account and using that account rather than the user's to locate the
desired bucket.  Existing code was already capable of parsing the account
out of the URL, but failed to use it when accessing the bucket.

This is part 3 of 4 fixes for tracker issue #22898.

Fixes: http://tracker.ceph.com/issues/22898
Signed-off-by: Marcus Watts <mwatts@redhat.com>
Conflicts:
 The master branch has entirely different code for this logic
 which already provides similar functionality.