client: add cap_dirtier_uid/gid to CapSnap
Set new fields to whatever the inode has at the time that we create the
CapSnap. In the event that it was unset at the time of the creation and
becomes dirty afterward, we set it again at finish.

This ensures that we send a valid uid/gid in the event that it had dirty
caps at any point during that period.

Tracker: http://tracker.ceph.com/issues/22801
Signed-off-by: Jeff Layton <jlayton@redhat.com>