client: add new ceph_mount_perms_set function
Allow programs to craft a UserPerm structure to use as the default
perms on the mount. The UserPerm is copied, so the caller can delete
the model UserPerm once the call is complete.

Note that we do not currently allow setting this once ceph_mount has
been called, as there are potential races that could occur while copying
it. We could eventually allow that, but we'd have to ensure that the
default_perms remain valid for accessors operating during and after the
copy.

Tracker: http://tracker.ceph.com/issues/22802
Signed-off-by: Jeff Layton <jlayton@redhat.com>