mgr/influx: Set a default retention of 8 weeks on database
Without a retention policy it is up to the administrator to set
a retention policy on the data.

This can easily be forgotten resulting in a ever growing InfluxDB
database.

Users can change the retention policy if they want to afterwards,
this is just the default policy set by the InfluxDB module when
it initially creates the database.

Signed-off-by: Wido den Hollander <wido@42on.com>