Reject invalid swift acls.
Previous versions silently ate invalid acls.  Most users
found this extremely confusing, and this behavior is not
consistent with the native openstack swift implementation.

There was existing logic that was supposed to fail under extreme
circumstances (such as no memory), this logic was not functional.
Logic to check the result from RGWAccessControlPolicy_SWIFT::create()
could never trip because "boolean" true is 1, not <0.

With this revision, invalid acls are now rejected.  This change
also makes the existing error case logic functional.

Fixes: http://tracker.ceph.com/issues/22896
Signed-off-by: Marcus Watts <mwatts@redhat.com>
Conflicts:
 this change does not apply to master; code and behavior are
 both quite different there.