rpm: adjust ceph-{osdomap,kvstore,monstore}-tool feature move
in pre luminous era, ceph-test "Requires" ceph-common without specifying
any version requirement. this allows a luminous ceph-common package to
co-exist with a kraken ceph-test, so, even if we let luminous ceph-osd
to "Require" the ceph-base of the same version, we are not able to
address following error:

Transaction check error:
  file /usr/bin/ceph-kvstore-tool from install of ceph-base-2:12.2.2-773.g3d84e35.el7.x86_64 conflicts with file from package ceph-test-1:11.2.1-12.gad30823.el7.x86_64

so, per [1], we can use Conflict in this case, as ceph-{base,osd,mon} do
not depend on ceph-test. they are installable independently of whether
ceph-test is installed.

[1] https://fedoraproject.org/wiki/Packaging:Conflicts#Splitting_Packages

Fixes: http://tracker.ceph.com/issues/22558
Signed-off-by: Kefu Chai <kchai@redhat.com>