tests: client-upgrade-kraken: upgrade ceph-test, too
Upgrade test was failing because, in luminous and above, three files were moved
from ceph-test to ceph-{base,mon,osd}. When the latter packages are upgraded,
they conflict with ceph-test, which was not being upgraded.

Since it is no longer necessary to avoid upgrading ceph-test, and since
b07aa210aa0ede54ffc3dbe49e334bd51a8f6342 tied ceph-test to the ceph-common of
the exact same version, this bug could be fixed by removing ceph-test from
the list of packages to be excluded from the upgrade.

Fixes: https://tracker.ceph.com/issues/22558
Signed-off-by: Nathan Cutler <ncutler@suse.com>