mds: allow storing open file table in multiple omaps
Adjust number of objects to store open file table according to size of
the open file table. The basic idea is, when saving an entry for the
first time, assign it to an omap object. If all existing omap objects
are too full, allocate an new object. Total number of objects is stored
first object's omap header.

To guarantee consistency of open file table in the case of updating
multiple objects. mds needs to first journal the updates, then update
the original omaps.

Signed-off-by: "Yan, Zheng" <zyan@redhat.com>