mgr: mgr_stats_period from 5 -> 2
This gives a more responsive experience for 'ceph -s' and for the
dashboard.  So far we have not seen significant load on the mgr daemon on
even largish clusters.

Lower the default for now.  We can either make the system auto-tune this
value if the mgr is loady, or we can simply ask large cluster operators to
increase it if the mgr is using too much CPU.

Signed-off-by: Sage Weil <sage@redhat.com>