rbd: fix thread_offsets calculation of rbd bench
This patch fixes the calculation of the thread_offset vector
for sequential I/O of rbd bench command.

The rbd bench command doesn't access whole image of rbd,
because the some chunks are not assigned to threads.
This patch changes the way to calculate the thread_offsets
to assign all chunks to threads.

Signed-off-by: Hitoshi Kamei <hitoshi.kamei.xm@hitachi.com>
Cc: Mitsuo Hayasaka <mitsuo.hayasaka.hu@hitachi.com>