mgr/dashboard: fix frontend e2e tests
Updated protractor configuration: now it's possible to run the tests inside a
docker container, added params for the username and password and added login
during the prepare phase.

Chrome is required to run the e2e tests. This can be achieved by installing
Chrome in your system or by using a docker container with Chrome installed.

Changed some timer calls, so they are runned outside of angular. This was
preventing protractor from running the tests.

Signed-off-by: Tiago Melo <tmelo@suse.com>