cls/rgw: make usage_iterate_range()'s "truncated" parameter trustworthy
Set it to false whenever we identify that we've reached the end of our
range, even if the underlying OSD op says it could have given us more
results. Then rely on that instead of weird iter-based logic to tell
our client that it doesn't need to do more work.

Fixes: https://tracker.ceph.com/issues/22963

Signed-off-by: Greg Farnum <gfarnum@redhat.com>