mgr/dashboard_v2: fix and improve table details
There was a problem with the current implementation of the table details, when
compiled in production mode, where the method that created the details element
would throw an exception.

Besides fixing that, this commit also improves the way we define the details.
Now instead of sending the component name, through the 'detailsComponent' attr,
we need to get the current selection via 'updateSelection' attr and declare the
detail component inside the cdTable and add the cdTableDetail attr to that
element.

Signed-off-by: Tiago Melo <tmelo@suse.com>
(cherry picked from commit ab780bafc2a13be461e37407f5db8c602b7741d1)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>