rbd: rbd_discard_on_zero_write_same option support
Add the rbd_discard_on_zeroed_write_same option to
the writesame method of librdb, if applied.

If the rbd_discard_on_zero_write_same option is true,
the data of rbd image are just discarded by writesame
when librdb receives the write operation of zero data.
Meanwhile, if the option is false, the zero data are
actually written to the image.

The default value of rbd_discard_on_zero_write_same
is true.

Signed-off-by: Hitoshi Kamei <hitoshi.kamei.xm@hitachi.com>
Cc: Mitsuo Hayasaka <mitsuo.hayasaka.hu@hitachi.com>