client: add ceph_ll_sync_inode
Ganesha's NFS SETATTR requests are expected to be synchronous, but
libcephfs can cache attribute changes if the client has the right caps.
An fsync after changing the attributes would flush the caps back to the
MDS, but we may not have an Fh in the context of a SETATTR request.

Add a way to sync a specific Inode, even when we don't have easy access
to a Fh.

Tracker: http://tracker.ceph.com/issues/23291
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 78628a5fde277927126d4d6e5d2957e7e6d42fb1)