common/bluestore: Lower bluestore_cache_size_ssd to 1GB
The default value of 3GB for bluestore_cache_size_ssd might be too
large for older systems migrating from FileStore to BlueStore.

It's very common to have a system with roughly 4GB of memory per OSD,
for example a system with 8 SSDs and 32GB of Memory.

This is not enough memory to safely run 8 OSDs with a cache size set
to 3GB.

To prevent users from having a bad experience when migrating to
BlueStore the default value is lowered to 1GB.

Admins can change this value to something matching their environment.

Signed-off-by: Wido den Hollander <wido@42on.com>