os/bluestore: avoid unexpected cache lru changes during trim
If we have an onode in the LRU that has a SharedBlob holding the last
reference to a collection, we can get a call chain like

_trim
        onode_map.remove()
                ~OnodeRef
                        ~SharedBlob
                                BufferSpace::_clear()
                                ~Collection
                                        ~OnodeSpace
                                                OnodeSpace::clear()
                                                        Cache::_rm_onode(...)

which can invalidate the _trim iterator p that is pointing to the item
right after the onode we're trimming.

Avoid this messed up situation by clearing the OnodeSpace when the
collection is removed.  (Note that we already check that no objects
exist, but we may still have negative Onodes (Onode::exists == false) in
the cache, causing the _trim problem above.)

Fixes: http://tracker.ceph.com/issues/21259
Signed-off-by: Sage Weil <sage@redhat.com>